utited users and usersList into users, add Redirect, refactor links on users in table
