+++ src/app.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
-import Users from './components/users'
+import { Switch, Route, Redirect } from 'react-router-dom'
+import Users from './layouts/users'
 import NavBar from './components/navBar'
-import { Switch, Route } from 'react-router-dom'
-import Main from './components/main'
-import Login from './components/login'
+import Main from './layouts/main'
+import Login from './layouts/login'
 import UserPage from './components/userPage'
 
 const App = () => {
@@ -15,6 +15,7 @@ const App = () => {
                 <Route path="/login" component={Login} />
                 <Route path="/users" exact component={Users} />
                 <Route path="/users/:userId?" component={UserPage} />
+                <Redirect to="/" />
             </Switch>
 
         </>
