+++ src/components/navBar.jsx
@@ -5,7 +5,7 @@ const NavBar = () => {
     return (
         <ul className="nav">
             <li className="nav-item">
-                <Link className="nav-link active" aria-current="page" to="/">Main</Link>
+                <Link className="nav-link" aria-current="page" to="/">Main</Link>
             </li>
             <li className="nav-item">
                 <Link className="nav-link" to="/login">Login</Link>
