+++ src/components/tableBody.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import _ from 'lodash'
-import { Link } from 'react-router-dom'
 
 const TableBody = ({ data, columns }) => {
     const renderContent = (item, column) => {
@@ -22,11 +21,7 @@ const TableBody = ({ data, columns }) => {
                 <tr key={item._id}>
                     {Object.keys(columns).map(column => (
                         <td key={column}>
-                            { column === 'name'
-                                ? <Link to={`/users/${item._id}`} >
-                                    {item.name}
-                                </Link>
-                                : renderContent(item, column) }
+                            { renderContent(item, column) }
                         </td>
                     ))}
                 </tr>
