+++ src/components/usersTable.jsx
@@ -3,12 +3,16 @@ import PropTypes from 'prop-types'
 import BookMark from './bookmark'
 import QualitiesList from './qualitiesList'
 import Table from './table'
+import { Link } from 'react-router-dom'
 
 const UsersTable = ({ users, onDelete, onToggleBookMark, onSort, selectedSort }) => {
     const columns = {
         name: {
             path: 'name',
-            name: 'Имя'
+            name: 'Имя',
+            component: (user) => (
+                <Link to={`/users/${user._id}`}>{ user.name }</Link>
+            )
         },
         qualities: {
             name: 'Качества',
